requirements.txt
certifi==2020.12.5
numpy==1.20.2
Pillow==8.2.0
pthflops
torch==1.8.1
torchaudio==0.8.1
torchvision==0.9.1
typing-extensions==3.10.0.0
absl-py==0.12.0
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
dataclasses
flatbuffers==1.12
gast==0.3.3
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==4.0.1
Keras==2.4.3
keras-flops==0.1.2
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy==1.18.5
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.2.52
opt-einsum==3.3.0
Pillow
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.4.1
six
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.0
termcolor==1.1.0
torch==1.7.1
torchvision==0.8.2
typing-extensions
urllib3==1.26.4
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.1